lint
SCCS-vsn: sys/kern/kern_clock.c 4.23
SCCS-vsn: sys/vax/uba/ts.c 4.18
SCCS-vsn: sys/vax/uba/lp.c 4.19
SCCS-vsn: sys/vax/uba/tm.c 4.40
SCCS-vsn: sys/vax/mba/hp.c 4.41
SCCS-vsn: sys/vax/uba/ts.c 4.19
SCCS-vsn: sys/vax/vax/trap.c 4.10

+++ usr/src/sys/vax/uba/tm.c
@@ -1,4 +1,4 @@
-/*	tm.c	4.39	81/07/05	*/
+/*	tm.c	4.40	81/07/09	*/
 
 #include "te.h"
 #include "ts.h"
@@ -428,7 +428,7 @@ loop:
 			sc->sc_timo = 5 * 60;
 		} else {
 			um->um_tab.b_active = SCOM;
-			sc->sc_timo = imin(imax(10 * bp->b_repcnt, 60), 5*60);
+			sc->sc_timo = imin(imax(10*(int)bp->b_repcnt,60),5*60);
 		}
 		if (bp->b_command == TM_SFORW || bp->b_command == TM_SREV)
 			addr->tmbc = bp->b_repcnt;
